Implement extraction of metadata from video and image files
This allows us to get timestamps, and also the data about gps
coordinates and duration out of the video or image files themselves

To do so we use MP4 parsing, and EXIF data parsing.

Also put in an '+' add button to add files without dragging and
dropping to a specific location.